pyproject.toml
[tool.poetry]
name = "single_jobboard"
version = "0.1.0"
description = ""
authors = ["Son Thanh Nguyen"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "^3.141.0"
webdriver_manager = "^2.4.0"
rich = "^1.0.3"
starlette = "^0.13.4"
uvicorn = "^0.11.5"
attrs = "^19.3.0"
arsenic = "^19.1"
ujson = "^2.0.3"
bs4 = "^0.0.1"
lxml = "^4.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
